@@ -1,9 +1,6 @@
-const autoprefixer = require("autoprefixer");
-const tailwindcss = require("tailwindcss");
-
 module.exports = {
-    plugins: [
-        require("tailwindcss"),
-        require("autoprefixer"),
-    ]
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
 }
