@@ -1,8 +1,11 @@
 module.exports = {
-    purge: [
-      "./src/**/*.vue",
-    ],
-    theme: {},
-    variants: {},
-    plugins: [],
-  }
+  purge: ['./index.html', './src/**/*.{vue,js,ts,jsx,tsx}'],
+  darkMode: false, // or 'media' or 'class'
+  theme: {
+    extend: {},
+  },
+  variants: {
+    extend: {},
+  },
+  plugins: [],
+}
